Add event listener to index.js and refactor API logic
