+++ src/index.js
@@ -1,3 +1,11 @@
-/* eslint-disable no-unused-vars */
-import page from './dom.js';
 import external from './api.js';
+
+const form = document.getElementById('form');
+form.addEventListener('submit', (e) => {
+  e.preventDefault();
+  const data = Object.fromEntries(new FormData(form));
+  const city = data['city'];
+  const country = data['country'];
+  external.getWeather(city, country);
+  form.reset();
+});
